third_party/WebKit/LayoutTests/webmidi/requestmidiaccess-expected.txt

Tests navigator.requestMIDIAccess.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS requestMIDIAccess() succeeded with access [object MIDIAccess].
PASS access.inputs().length is 1
PASS access.outputs().length is 1
PASS input attributes are correct
PASS output attributes are correct
PASS navigator.requestMIDIAccess() did not throw exception.
PASS navigator.requestMIDIAccess() did not throw exception.
PASS requestMIDIAccess() was correctly blocked for System Exclusive access with error [object DOMError].
PASS successfullyParsed is true

TEST COMPLETE
